Changed set up of code in AddBookForSaleActivity for clarity.
RequestBookActivity was not working and app crashed everytime when opened => Fixed that.

Changed Image in both "Add Book for Sale" and "Request a Book" to better guide users. Picture now says "Please upload image" and if user tries to Submit book without uploading Image, the app no longer crashes and the image changes to "No Image". (But on to-do list is changing that picture to say "NO IMAGE UPLOADED therefore NO BOOK WAS ADDED" or something similar).